blockchain v0: add TestReactor_AbruptDisconnect
